DataLoader.CSV

Documentation

parseCSV :: (CSVParser a, Contextable a) => ByteString -> Int -> DataSet a

parseCSV : Can parse a contextable type generating a dataset of this type. The second parameter, integer, defines the length of the set of codes. However, only length one and two are implemented (entities and simple relations). Therefore, if the integer is other than 1 or 2, an EntitySet will be generated as it is the default one.

parseCSVstruc :: ByteString -> [(String, [String])]

parseCSVstruc : Function to load a structure of regions, as it is provided by the World Bank