Maps.DataType

Documentation

class Contextable a where

class Contextable : Type class for all the types which can be inserted in a dataset and, therefore, for which a context can be created. it contains a method

Minimal complete definition

autoContext

Methods

generateDataset :: [(Code, a)] -> DataSet a

autoContext : Function that generates a predefined dataset from a list of tuples (Code, a), being a Contextable

generateRelationset :: [(Code, Code, a)] -> DataSet a

autoContext : Function that generates a predefined dataset from a list of triples (Code, Code, a), being a Contextable

data DataSet a

The polymorphic algebraic datatype DataSet has two constructors: EntitySet refers to data about entities, in this case regions. It contains a parameter v, which is a hash map with the data ordered by codes y a context RelationSet refers to data about relations, in this case links or flows. It contains a parameter rv, which is a list of triples containing the connected nodes and the value. It also has a context parameter.

Constructors

EntitySet

Fields

v :: Map Code a

c :: Context

RelationSet

Fields

rv :: [(Code, Code, a)]

c :: Context

data Context

The Context algebraic datatype provides information relevan to the data contained in the dataset, such as the maximum, the nominal values... It has a constructor for every kind of data, namely Nominal, Ordinal and Ratio, and a predefined brewerset.

Constructors

NominalContext

Fields

values :: [StvNominal]

brewerset :: ColorCat

OrdinalContext

Fields

mx :: StvOrdinal

mn :: StvOrdinal

step :: StvOrdinal

brewerset :: ColorCat

RatioContext

Fields

rmx :: StvRatio

rmn :: StvRatio

brewerset :: ColorCat

type StvOrdinal = Int

type StvNominal = String

DATATYPE MODULE ------------------------------------------------------------------------------------------------- ................................................................................................................... ___________________________________________________________________________________________________________________

Synonims for the different kinds of data, and the code. This helps to mantain the code more legible.

type StvRatio = Double

class RetHue a where

The Type Classes for the retinal variables. If a type implements a retinal type class that means that it is appropriate representate it through it. For example a nominal type implements hue but not size.

class RetHue : Interface for the retinal variable hue, also referred as color in the literature. it contains a method hue, receiving a context and a (Maybe) a value for which it will return a color (or a default one if it receives nothing) the method hue' receives an extra parameter explicitely indicating the range of values, which is calculated automatically in the previous one.

Minimal complete definition

hue, hue'

Methods

hue :: Context -> Maybe a -> Kolor

class RetLightless a where

class RetLightless : Interface for the retinal variable ligthless or opacity. it contains a method lightless, receiving a context and a (Maybe) a value for which it will return a size (or a default one if it receives nothing)

Methods

lightless :: Context -> Maybe a -> Double

class RetSize a where

class RetSize : Interface for the retinal variable size. it contains a method size, receiving a context and a (Maybe) a value for which it will return a size (or a default one if it receives nothing) the method size' receives an extra parameter explicitely indicating the range of values, which is calculated automatically in the previous one.

Minimal complete definition

size, size'

Methods

size' :: (Double, Double) -> Context -> Maybe a -> Double

dsToNominal :: DataSet a -> (a -> Int) -> [StvNominal] -> DataSet StvNominal

dsToNominal : This function transforms a DataSet from one type to a Nominal. We provide The original dataset and a function that maps values to Int, ans a set of possible values

dsTransform :: Contextable b => DataSet a -> (a -> b) -> DataSet b

dsTransform : This function transforms a DataSet from one type to another. We provide The original dataset and a function that maps values of the type a to values of the type b. B has to be a Contectable type, that means, either StvNominal, StvOrdinal or StvRatio at this point.

compositeDs :: (Contextable c, Eq b) => (a -> b -> c) -> DataSet a -> DataSet b -> DataSet c

compositeDs : This function combines two datasets into one, through a provided function. We provide a function that receives one param of each dataset, and those two datasets. C has to be a Contectable type, that means, either StvNominal, StvOrdinal or StvRatio at this point. This function only works for a couple of EntitySets or a couple of RelationSets. If a mix is provided, an empty Dataset will be returned.