Maps.Map

Documentation

buildDiagram :: [DiagramList] -> Diagram B

buildDiagram : Symply merges a list of diagrams in absolute positions and generates a resulting composition. The pattern for building diagrams is the following: map = buildDiagram $ [printRegions fAll] from ($ tree) Here we can apply all the mapData func data that we want: buildDiagram $ [printRegions mapData mapLightless ds2 # fAll] from ($ tree) If we want to map things over another set of regions we add it to the array: map = buildDiagram $ [printRegions fSub 1, printRegions fSub 2] from ($ tree) If we want to arrows over regions we add them as well to the array: map = buildDiagram $ [printArrows dsArrows mapData mapLightless ds2 # fSub 2] from ($ tree)

fAll :: (Regions -> DiagramList) -> SPtree -> DiagramList

fAll : This function receives a function and a tree, and maps the function to the lower level (Level of detail 0) of the tree

fSub :: Int -> (Regions -> DiagramList) -> SPtree -> DiagramList

fSub : This function receives a function and a tree, and maps the function to the desired level (int param) of the tree

from :: [SPtree -> DiagramList] -> ((SPtree -> DiagramList) -> DiagramList) -> [DiagramList]

from : This function receives a list of functions and maps it to a tree, to produce a diagram list that will be used in buildDiagram

printRegions :: Regions -> DiagramList

printRegions : Generates a list of Diagrams from a selection of regions

printArrows :: RetSize a => DataSet a -> Regions -> DiagramList

printArrows : Generates a list of arrows from a dataset and a

mapData :: (DataSet a -> RegDiagram -> RegDiagram) -> DataSet a -> (Regions -> DiagramList) -> Regions -> DiagramList

mapData : This function receives: 1- A function that applies data to a diagram and returns it updated 2- A dataset with the data to apply 3- A function that generates diagrams from regions 4- The regions The function generates the diagrams with the regions(4) and the functions(3), and to every diagram created applies the transforming function (1) with the data (2)

mapHue :: RetHue a => DataSet a -> RegDiagram -> RegDiagram

mapHue : Applies a hue to a diagram

mapLightless :: RetLightless a => DataSet a -> RegDiagram -> RegDiagram

mapLightless : Applies ligntless to a diagram