Maps.Spatial

Documentation

data Region

The Region is an Algebraic Data Type that contains the class or identifier of a shape and a list of Diagram's paths with its points. A path is a set of located trails.

Constructors

Base

Fields

code :: Code

coor :: [Path V2 Double]

Group

Fields

code :: Code

level :: Int

nested :: [(String, Region)]

ids :: [String]

type SPtree = Seq [(Code, Region)]

The synonim SPtree represents the spatial structure. It is a sequence of lists, every item of the sequence corresponds to one level of detail (the fact that it is indexed motivates the use of a sequence instead of a list). Inside every level we have a list of tuples with the regions belonging to that level, and their codes. The base level is always the first (index 0), and the rest are constructed over it

Level 0: Spain | Portugal | Germany | ...
                         
Level 1:  Iberian Pen     /
                        /  ...
Level 2:          Europe

rcenter :: Code -> [Region] -> Maybe (P2 Double)

rcenter : Calculates the center of a group of regions satisfying a code.

buildTree :: [Region] -> SPtree

buildTree : Initializes a tree with a single level and the provided regions in the floor.

mostSpecific :: SPtree -> [Region]

mostSpecific : Retrieves the regions from the most specific level of detail, that means, level 0

addHierarchy :: [(String, [String])] -> Int -> SPtree -> SPtree

addHierarchy : Loads a complete hierarchy, specified through tuples of (region, list of nested regions)

area :: Region -> SPtree -> [Path V2 Double]

area : Returns the area of a region, which is that contained within the points of coor for a Base region or the accumulation of the areas of the nested Regions. This is a recursive function.

lookup :: Code -> SPtree -> Maybe Region

lookup : Returns the region of the code provided or nothing if it's not in the tree.